requirements/test.txt
-r base.txt
coverage==4.4.1           # via pytest-cov
py==1.4.34                # via pytest
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.1.2
